dailybuild: 64bit 構成時の warning 抑止
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7814 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Physics/PHOpGroup.cpp
@@ -31,7 +31,7 @@ void PHOpGroup::delPInds(int gPIndex)
 
 	}
 	gPInd.swap(newgpinds);
-	gNptcl = gPInd.size();
+	gNptcl = (int)gPInd.size();
 }
